# internal/docs/swagger.yaml
definitions:
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  handlers.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  models.Reminder:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_sent:
        type: boolean
      message:
        type: string
      send_at:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
info:
  contact: {}
paths:
  /reminders:
    get:
      consumes:
      - application/json
      description: Получить список всех напоминаний
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Reminder'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получение всех напоминаний
      tags:
      - reminders
    post:
      consumes:
      - application/json
      description: Создать новое напоминание с предоставленными деталями
      parameters:
      - description: Reminder object
        in: body
        name: reminder
        required: true
        schema:
          $ref: '#/definitions/models.Reminder'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Reminder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Создать напоминание
      tags:
      - reminders
  /reminders/{id}:
    delete:
      consumes:
      - application/json
      description: Удалить напоминание по идентификатору, если оно не было отправлено
      parameters:
      - description: Reminder ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Удалить напоминание
      tags:
      - reminders
    put:
      consumes:
      - application/json
      description: Обновить напоминание с указанным идентификатором
      parameters:
      - description: Reminder ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated reminder object
        in: body
        name: reminder
        required: true
        schema:
          $ref: '#/definitions/models.Reminder'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Reminder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Обновление существующего напоминания
      tags:
      - reminders
  /reminders/{user_id}:
    get:
      consumes:
      - application/json
      description: Получить все напоминания для конкретного пользователя
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Reminder'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Поиск напоминаний по user_id
      tags:
      - reminders
swagger: "2.0"